Initial release
This release contains all code, notebooks, examples, and assets created during the last four years.
We removed all client related code and/or examples and did some cleanup of the codebase. The different
parts of the repository are in different states of development but most of them where already used in
educational and training sessions with participants from a wide range of industries.

Feedback and contributions are welcome